# apis/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "description": "Get a products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get a products",
                "parameters": [
                    {
                        "description": "GetProductsRequest",
                        "name": "GetProductsRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.GetProductsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Create a new product",
                "parameters": [
                    {
                        "description": "CreateProductRequest",
                        "name": "CreateProductRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/services.CreateProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "description": "Get a product by the ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get a product by the ID",
                "parameters": [
                    {
                        "description": "GetProductRequestById",
                        "name": "GetProductRequestById",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.GetProductRequestById"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.GetProductByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update product with product info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Update product with product info",
                "parameters": [
                    {
                        "description": "UpdateProductRequest",
                        "name": "UpdateProductRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/services.UpdateProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a product by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Delete a product by id",
                "parameters": [
                    {
                        "description": "DeleteProductRequest",
                        "name": "DeleteProductRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.DeleteProductRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/services.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "enums.CategoryType": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10
            ],
            "x-enum-varnames": [
                "Electronics",
                "Fashion",
                "Beauty",
                "Sports",
                "Automotive",
                "Health",
                "Beverages",
                "Books",
                "Toys",
                "Home"
            ]
        },
...
        "enums.ProductStatus": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "Active",
                "InActive",
                "Discontinued"
            ]
        },
...
        "services.CreateProductRequest": {
            "type": "object",
            "required": [
                "categoryId",
                "description",
                "imageUrl",
                "name",
                "price",
                "stock"
            ],
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/enums.ProductStatus"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "services.CreateProductResponse": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "string"
                }
            }
        },
...
        "services.DeleteProductRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "services.GetProductByIdResponse": {
            "type": "object",
            "properties": {
                "categoryDescription": {
                    "type": "string"
                },
                "categoryName": {
                    "type": "string"
                },
                "categoryType": {
                    "$ref": "#/definitions/enums.CategoryType"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/enums.ProductStatus"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "services.GetProductRequestById": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "services.GetProductResponse": {
            "type": "object",
            "properties": {
                "categoryDescription": {
                    "type": "string"
                },
                "categoryName": {
                    "type": "string"
                },
                "categoryType": {
                    "$ref": "#/definitions/enums.CategoryType"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/enums.ProductStatus"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "services.GetProductsRequest": {
            "type": "object",
            "properties": {
                "order": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "pageLength": {
                    "type": "integer"
                },
                "sortBy": {
                    "type": "string"
                }
            }
        },
...
        "services.Result": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "perPage": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.GetProductResponse"
                    }
                },
                "requestedPageLength": {
                    "type": "integer"
                },
                "totalPage": {
                    "type": "integer"
                },
                "totalResults": {
                    "type": "integer"
                }
            }
        },
...
        "services.UpdateProductRequest": {
            "type": "object",
            "required": [
                "categoryId",
                "description",
                "imageUrl",
                "name",
                "price",
                "stock"
            ],
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/enums.ProductStatus"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "services.UpdateProductResponse": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "string"
                }
            }
        },
...
    }
}